refactor(frontend): PRレビュー指摘修正
- global.cssをfeatures/showcase/styles/からsrc/styles/に移動（cross-feature import解消）
- context/auth-context.tsxをcomponents/auth-provider.tsx + hooks/use-auth-context.tsに分離（bulletproof-react準拠）
