+++ frontend/src/app/providers/app-provider.tsx
@@ -2,7 +2,7 @@ import { TransportProvider } from "@connectrpc/connect-query";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import type { ReactNode } from "react";
-import { AuthProvider } from "../../features/auth/context/auth-context";
+import { AuthProvider } from "../../features/auth/components/auth-provider";
 import { queryClient } from "../../lib/query-client";
 import { transport } from "../../lib/transport";
