+++ frontend/src/features/auth/components/auth-context-internal.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import type { AuthContextValue } from "../types";
+
+export const AuthContext = createContext<AuthContextValue | null>(null);
