+++ frontend/src/features/auth/components/require-auth.tsx
@@ -1,5 +1,5 @@
 import { Navigate } from "react-router";
-import { useAuthContext } from "../context/auth-context";
+import { useAuthContext } from "../hooks/use-auth-context";
 
 export function RequireAuth({ children }: { children: React.ReactNode }) {
   const { isAuthenticated, isLoading } = useAuthContext();
